requirements.txt
#pip install -r requirements.txt

python-binance==1.0.17
requests==2.31.0
urllib3==2.2.1
questionary==2.0.1
